editorial: change IfAbruptCloseAsyncIterator location
The “Iterator Abstract Operations” subsection of “Iteration” from the
old version of proposal-iterator-helpers and
proposal-async-iterator-helpers is obsolete.

This means that the new IfAbruptCloseAsyncIterator operation (shared with proposal-async-iterator-helpers) no longer has a place to
live.

The ecma262 spec placed
proposal-iterator-helpers’ analogous IfAbruptCloseIterator operation
into a “Operations on Iterator Objects” subsection of “Abstract Operations”.

This commit places IfAbruptCloseAsyncIterator into an analogous
“Operations on Async Iterator Objects” subsection…although Async
Iterator Objects do not yet exist!

The editors may choose to name this new subsection whatever they want.
Or they might elect to not make a new subsection and foist
IfAbruptCloseAsyncIterator on the existing “Operations on Iterator
Objects” subsection instead, until proposal-async-iterator-helpers makes
Stage 4.

This is an editorial change.